chore(editor): Improve VS Code extension dev experience (#16087)
Improve the experience of launching the VS Code extension for development purposes. Now it will build it for you automatically when launching the editor extension, rather than you needing to build it yourself.

We could also maybe keep a launch setup that doesn't do this alongside this one, if others prefer using that workflow.

See this video for the workflow of building+launching this from the debug menu in VS Code:

https://github.com/user-attachments/assets/738368d4-88fd-424b-ab96-4b1cfed38491